vSomeIP 1.3.0

+++ implementation/service_discovery/src/fsm_base.cpp
@@ -1,5 +1,4 @@
-// Copyright (C) 2014 BMW Group
-// Author: Lutz Bichler (lutz.bichler@bmw.de)
+// Copyright (C) 2014-2015 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -9,31 +8,27 @@
 namespace vsomeip {
 namespace sd {
 
-fsm_base::fsm_base(boost::asio::io_service &_io) : timer_(_io) {
+fsm_base::fsm_base(boost::asio::io_service &_io)
+        : timer_(_io) {
 }
 
 fsm_base::~fsm_base() {
 }
 
 void fsm_base::start_timer(uint32_t _milliseconds) {
-	timer_.expires_from_now(std::chrono::milliseconds(_milliseconds));
-	timer_.async_wait(
-		std::bind(
-			&fsm_base::timer_expired,
-			shared_from_this(),
-			std::placeholders::_1
-		)
-	);
+    timer_.expires_from_now(std::chrono::milliseconds(_milliseconds));
+    timer_.async_wait(
+            std::bind(&fsm_base::timer_expired, shared_from_this(),
+                    std::placeholders::_1));
 }
 
 void fsm_base::stop_timer() {
-	timer_.cancel();
+    timer_.cancel();
 }
 
 uint32_t fsm_base::expired_from_now() {
-	return std::chrono::duration_cast<
-				std::chrono::milliseconds>(
-					timer_.expires_from_now()).count();
+    return (uint32_t) std::chrono::duration_cast < std::chrono::milliseconds
+            > (timer_.expires_from_now()).count();
 }
 
 } // namespace sd
